web-control/requirements.txt
Flask==2.3.3
PyYAML==6.0.1
requests==2.31.0
Werkzeug==2.3.7
